+++ src/placer/globalPlacer/GPSimPL/boxbin.cc
@@ -69,10 +69,10 @@ void BoxBin::update_cell_area_white_space(std::vector<std::vector<GridBin> > &gr
       total_cell_area += grid_bin_matrix[x][y].cell_area;
     }
   }
-  filling_rate = total_cell_area/(double)total_white_space;
+  filling_rate = double(total_cell_area)/double(total_white_space);
 }
 
-void BoxBin::UpdateCellAreaWhiteSpaceFillingRate(std::vector<std::vector<int > > &grid_bin_white_space_LUT, std::vector<std::vector<GridBin> > &grid_bin_matrix) {
+void BoxBin::UpdateCellAreaWhiteSpaceFillingRate(std::vector<std::vector<unsigned long int>> &grid_bin_white_space_LUT, std::vector<std::vector<GridBin> > &grid_bin_matrix) {
   if (ll_index.x == 0) {
     if (ll_index.y == 0) {
       total_white_space = grid_bin_white_space_LUT[ur_index.x][ur_index.y];
@@ -97,7 +97,7 @@ void BoxBin::UpdateCellAreaWhiteSpaceFillingRate(std::vector<std::vector<int > >
       total_cell_area += grid_bin_matrix[x][y].cell_area;
     }
   }
-  filling_rate = total_cell_area/(double)total_white_space;
+  filling_rate = double(total_cell_area)/double(total_white_space);
 }
 
 void BoxBin::ExpandBox(int grid_cnt_x, int grid_cnt_y) {
@@ -294,8 +294,8 @@ bool BoxBin::write_cell_in_box(std::string const &NameOfFile, std::vector<Block>
   return true;
 }
 
-double BoxBin::white_space_LUT(std::vector< std::vector< int > > &grid_bin_white_space_LUT, GridBinIndex &ll, GridBinIndex &ur) {
-  double white_space;
+unsigned long int BoxBin::white_space_LUT(std::vector< std::vector<unsigned long int>> &grid_bin_white_space_LUT, GridBinIndex &ll, GridBinIndex &ur) {
+  unsigned long int white_space;
   if (ll.x == 0) {
     if (ll.y == 0) {
       white_space = grid_bin_white_space_LUT[ur.x][ur.y];
@@ -317,9 +317,9 @@ double BoxBin::white_space_LUT(std::vector< std::vector< int > > &grid_bin_white
   return white_space;
 }
 
-bool BoxBin::update_cut_index_white_space(std::vector< std::vector< int > > &grid_bin_white_space_LUT) {
+bool BoxBin::update_cut_index_white_space(std::vector< std::vector<unsigned long int>> &grid_bin_white_space_LUT) {
   double error, minimum_error = 1;
-  double white_space_low;
+  unsigned long int white_space_low;
   int index_give_minimum_error;
   if (cut_direction_x) {
     if (ll_index.y == ur_index.y) return false;
@@ -331,9 +331,9 @@ bool BoxBin::update_cut_index_white_space(std::vector< std::vector< int > > &gri
     for (cut_ur_index.y = ll_index.y; cut_ur_index.y < ur_index.y-1; cut_ur_index.y++) {
       //std::cout << cut_ur_index.y << "\n";
       white_space_low = white_space_LUT(grid_bin_white_space_LUT, ll_index, cut_ur_index);
-      error = std::fabs(white_space_low/total_white_space-0.5);
+      error = std::fabs(double(white_space_low)/double(total_white_space)-0.5);
       if (error < minimum_error) index_give_minimum_error = cut_ur_index.y;
-      if (white_space_low/total_white_space > 0.5) break;
+      if (double(white_space_low)/double(total_white_space) > 0.5) break;
     }
     if (cut_ur_index.y != index_give_minimum_error) {
       cut_ur_index.y = index_give_minimum_error;
@@ -351,9 +351,9 @@ bool BoxBin::update_cut_index_white_space(std::vector< std::vector< int > > &gri
     for (cut_ur_index.x = ll_index.x; cut_ur_index.x < ur_index.x-1; cut_ur_index.x++) {
       //std::cout << cut_ur_index.x << "\n";
       white_space_low = white_space_LUT(grid_bin_white_space_LUT, ll_index, cut_ur_index);
-      error = std::fabs(white_space_low/total_white_space-0.5);
+      error = std::fabs(double(white_space_low)/double(total_white_space)-0.5);
       if (error < minimum_error) index_give_minimum_error = cut_ur_index.x;
-      if (white_space_low/total_white_space > 0.5) break;
+      if (double(white_space_low)/double(total_white_space) > 0.5) break;
     }
     if (cut_ur_index.x!= index_give_minimum_error) {
       cut_ur_index.x = index_give_minimum_error;
@@ -364,12 +364,14 @@ bool BoxBin::update_cut_index_white_space(std::vector< std::vector< int > > &gri
   }
 }
 
-bool BoxBin::update_cut_point_cell_list_low_high(std::vector<Block> &Nodelist, long int &box1_total_white_space, long int &box2_total_white_space) {
+bool BoxBin::update_cut_point_cell_list_low_high(std::vector<Block> &Nodelist, unsigned long int &box1_total_white_space, unsigned long int &box2_total_white_space) {
   // this member function will be called only when two white spaces are not different from each other for several magnitudes
-  int cell_area_low = 0;
+  unsigned long int cell_area_low = 0;
   double cut_line_low, cut_line_high;
   double cut_line = 0;
-  double ratio = 1 + box2_total_white_space/(double)box1_total_white_space;
+  double ratio = 1 + double(box2_total_white_space)/double(box1_total_white_space);
+  //double tmp_ratio = 0;
+  //TODO: this method can be accelerated
   Block *node;
   if (cut_direction_x) {
     cut_ur_point.x = ur_point.x;
@@ -387,6 +389,8 @@ bool BoxBin::update_cut_point_cell_list_low_high(std::vector<Block> &Nodelist, l
         }
       }
       //std::cout << cell_area_low/(double)total_cell_area << "\n";
+      //tmp_ratio = double(total_cell_area)/double(cell_area_low);
+      //TODO: this precision has some influence on the final result
       if (cell_area_low > total_cell_area/ratio) {
         cut_line_high = cut_line;
       } else if (cell_area_low < total_cell_area/ratio) {
@@ -424,6 +428,7 @@ bool BoxBin::update_cut_point_cell_list_low_high(std::vector<Block> &Nodelist, l
         }
       }
       //std::cout << cell_area_low/(double)total_cell_area << "\n";
+      //tmp_ratio = double(total_cell_area)/double(cell_area_low);
       if (cell_area_low > total_cell_area/ratio) {
         cut_line_high = cut_line;
       } else if (cell_area_low < total_cell_area/ratio) {
@@ -450,7 +455,7 @@ bool BoxBin::update_cut_point_cell_list_low_high(std::vector<Block> &Nodelist, l
 }
 
 bool BoxBin::update_cut_point_cell_list_low_high_leaf(std::vector<Block> &Nodelist, int &cut_line_w, int ave_blk_height) {
-  int cell_area_low = 0;
+  unsigned long int cell_area_low = 0;
   double cut_line = 0;
   // the above three cut-lines are for cells, thus they needs to be double
   double ratio = 2.0;
@@ -503,12 +508,12 @@ bool BoxBin::update_cut_point_cell_list_low_high_leaf(std::vector<Block> &Nodeli
     }
 
     /* and then, find the index of cell, the total cell area below which is closest to one half of the total cell area */
-    int tmp_tot_cell_area_low = 0;
+    unsigned long int tmp_tot_cell_area_low = 0;
     int index_tot_cell_low_closest_to_half = 0;
     for (size_t i=0; i<cell_list.size(); i++) {
       node = &Nodelist[cell_list[i]];
       tmp_tot_cell_area_low += node->Area();
-      cell_area_low_percentage = tmp_tot_cell_area_low/(double)total_cell_area;
+      cell_area_low_percentage = double(tmp_tot_cell_area_low)/double(total_cell_area);
       //std::cout << i << " " << cell_area_low_percentage << "\n";
       if (fabs(cell_area_low_percentage - low_white_space_total_ratio) < mini_error) {
         mini_error = fabs(cell_area_low_percentage - low_white_space_total_ratio);
@@ -566,12 +571,12 @@ bool BoxBin::update_cut_point_cell_list_low_high_leaf(std::vector<Block> &Nodeli
       cell_list[i] = tmp_cell_id;
     }
     /* second, find the index of cell, the total cell area below which is closest to one half of the total cell area */
-    int tmp_tot_cell_area_low = 0;
+    unsigned long int tmp_tot_cell_area_low = 0;
     int index_tot_cell_low_closest_to_half = 0;
     for (size_t i=0; i<cell_list.size(); i++) {
       node = &Nodelist[cell_list[i]];
       tmp_tot_cell_area_low += node->Area();
-      cell_area_low_percentage = tmp_tot_cell_area_low/(double)total_cell_area;
+      cell_area_low_percentage = double(tmp_tot_cell_area_low)/double(total_cell_area);
       //std::cout << i << " " << cell_area_low_percentage << "\n";
       if (fabs(cell_area_low_percentage - 1/ratio) < mini_error) {
         mini_error = fabs(cell_area_low_percentage - 1/ratio);
@@ -606,7 +611,7 @@ bool BoxBin::update_cut_point_cell_list_low_high_leaf(std::vector<Block> &Nodeli
     cut_ll_point.y = cut_line;
     cut_ur_point.y = cut_line;
     /* finally, the cut-line for white space is proportional to the total_cell_area_low */
-    cut_line_w = left + (int)((total_cell_area_low/(double)total_cell_area)*(right - left));
+    cut_line_w = left + (int)((double(total_cell_area_low)/double(total_cell_area))*(right - left));
   }
   return true;
 }
